+++ packages/react/src/core/pulp/pulpShared.ts
@@ -5,7 +5,7 @@ export function hasRenderedWithForceVisit(rendered: Pulp[] | null): boolean {
     if (!rendered) return false;
 
     for (const pulp of rendered) {
-        if (!(pulp instanceof HostTextPulp) && pulp.isForceToVisit) {
+        if (!(pulp instanceof HostTextPulp) && pulp.state.isForceToVisit) {
             return true;
         }
     }