+++ packages/react/src/core/pulp/pulpTypes.ts
@@ -1,7 +1,7 @@
-import {ReactElement} from 'react';
-import {ComponentPulp} from './ComponentPulp';
-import {HostComponentPulp} from './HostComponentPulp';
-import {HostTextPulp} from './HostTextPulp';
+import { ReactElement } from 'react';
+import { ComponentPulp } from './ComponentPulp';
+import { HostComponentPulp } from './HostComponentPulp';
+import { HostTextPulp } from './HostTextPulp';
 
 export enum PulpType {
   Component,
@@ -14,4 +14,9 @@ export type Pulp = HostTextPulp | HostComponentPulp | ComponentPulp;
 export interface SplitPulp<T extends ReactElement | string | number> {
   component: T | null;
   pulp: Pulp | null;
+}
+
+export interface PulpState {
+  readonly isSplitted: boolean;
+  readonly isForceToVisit: boolean;
 }